Move niNode:iterate to niAVObject
