+++ MWSE/NINodeLua.cpp
@@ -1,89 +1,13 @@
 #include "NINodeLua.h"
 
 #include "LuaManager.h"
-#include "LuaUtil.h"
-#include "StringUtil.h"
 
 #include "NIAVObject.h"
 #include "NIBillboardNode.h"
-#include "NIDefines.h"
 #include "NINode.h"
 #include "NISortAdjustNode.h"
 
-namespace {
-	bool passesTraverseFilters(const NI::AVObject* object, const std::unordered_set<unsigned int>& typeFilters, std::string_view prefix) {
-		bool passesFilter = typeFilters.empty() ? true : false;
-		if (!passesFilter) {
-			for (const auto type : typeFilters) {
-				if (object->isInstanceOfType((uintptr_t)type)) {
-					passesFilter = true;
-					break;
-				}
-			}
-		}
-		bool passesPrefix = prefix.empty() ? true : object->name && se::string::starts_with(object->name, prefix);
-		return passesFilter && passesPrefix;
-	}
-}
-
 namespace mwse::lua {
-	std::function<NI::Pointer<NI::AVObject>()> traverse(NI::Node* self, sol::optional<sol::table> param) {
-		bool recursive = getOptionalParam(param, "recursive", true);
-		std::string prefix = getOptionalParam(param, "prefix", std::string(""));
-		std::unordered_set<unsigned int> filters;
-
-		if (param) {
-			sol::table paramTable = param.value().as<sol::table>();
-			sol::object maybeValue = paramTable["type"];
-			if (maybeValue.valid()) {
-				if (maybeValue.is<unsigned int>()) {
-					filters.insert(maybeValue.as<unsigned int>());
-				}
-				else if (maybeValue.is<sol::table>()) {
-					sol::table filterTable = maybeValue.as<sol::table>();
-					for (auto [_, value] : filterTable) {
-						filters.insert(value.as<unsigned int>());
-					}
-				}
-				else {
-					throw std::invalid_argument("Iteration can only be filtered by a NI object type, or a table of object types.");
-				}
-			}
-		}
-
-
-		std::queue<NI::Pointer<NI::AVObject>> queue;
-		std::function<void(const NI::AVObject*)> traverseChild = [&](const NI::AVObject* object) {
-			if (!object->isInstanceOfType(NI::RTTIStaticPtr::NiNode)) {
-				return;
-			}
-
-			const auto asNode = static_cast<const NI::Node*>(object);
-			for (const auto& nodeChild : asNode->children) {
-				if (passesTraverseFilters(nodeChild, filters, prefix)) {
-					queue.push(nodeChild);
-				}
-				if (recursive) {
-					traverseChild(nodeChild);
-				}
-			}
-		};
-
-		if (passesTraverseFilters(self, filters, prefix)) {
-			queue.push(self);
-		}
-		traverseChild(self);
-
-		return [queue]() mutable -> NI::Pointer<NI::AVObject> {
-			if (queue.empty()) {
-				return nullptr;
-			}
-			auto ret = queue.front();
-			queue.pop();
-			return ret;
-		};
-	}
-
 	void bindNINode() {
 		// Get our lua state.
 		const auto stateHandle = LuaManager::getInstance().getThreadSafeStateHandle();
